backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
torch==2.1.1
torchvision==0.16.1
nibabel==5.1.0
pydicom==2.4.0
numpy==1.24.3
Pillow==10.2.0
minio==7.2.0
boto3==1.34.2
python-dotenv==1.0.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
httpx==0.25.2
falkordb==1.0.9
kaggle==1.6.14
pytest==7.4.3
pytest-asyncio==0.21.1
types-requests==2.31.0.10

# Rate Limiting
slowapi==0.1.9

# Observability: Prometheus metrics + distributed tracing
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-celery==0.43b0

# ONNX Runtime for optimized inference
onnx==1.15.0
onnxruntime==1.16.3
